remove subtree merge
I'm thinking that this was not a great idea to have included in the
first place. I don't really think it's helpful to have people using
subtree merges as a submodule replacement and I think it would be
confusing for the people that try. We can put this online somewhere as a
reference, but I no longer think it should be in the book.