maintain general structure for chapter summaries
The chapter summary for *10* is jarring due to prose inserted before the bold chapter number.  This patch attempts to maintain a similar pattern of the previous 9 chapter summaries removes this discord and maintains the original content. It may be that this difference is a means to call out the authors desire to read chapter 10 first. in that case maybe some other call out could be employed to prevent the jarring effect.