Improve subtree merge command line.
Subtree merge doesn't always work as intended here.

For example, I tried these instructions with two repositories,
repository A, which contained a README file, and repository B, which
contained README and AUTHORS files. I used read-tree to put repository
B into a subdirectory of repository A, then ran the git merge
incantation affected by this patch.

With Git 1.9.1 from the Ubuntu 14.04 repositories, git-merge decided
that the correct subtree of A to merge B into was the root directory
of A, perhaps because the directory structure was similar enough to
satisfy some heuristic.

Replace the subtree strategy with recursive and explicit hint about
where the subtree is, which worked as intended. The output ("Squash
commit [...] as requested") was unchanged.

Signed-off-by: Marti Bolivar <mbolivarmullen@gmail.com>