fixed error in the imap section
`git send-email` does not put the patches into an IMAP folder, but the
command `git imap-send`.
I've changed the command and added the output that occurs when using the
imap-send command. I also wrote a short sentence to make use of the previous
output used in the document, plus I added an example for an SMTP config.

Added `git imap-send` to the section email of git commands.