Bold and italic follow Asciidoc semantics
The book used a combination of Markdown and Asciidoc syntax for
bold and italic text. In the documentation, it says that * are
used for highlighting the whole word or phrase, and ** are used
to highlight a part of the word. The render still worked fine
before, but the source-code now follows the semantics.

http://asciidoctor.org/docs/user-manual/#bold-and-italic