Solve authentication problem
There is an issue with the authentication section. The initial ref
advertisement request will fail without even giving an opportunity for
authentication. This is described in the git-http-backend man page
examples.

Update authentication section according to the example to fix the problem.
Merge the section with the git-core access authorization section for
brevity.

Signed-off-by: Vitaly Kuznetsov <vitalyq@users.noreply.github.com>