Revert wrong fix about --no-merges
There was a fix in 6d9867ff2394cb8a08fa3e013b082a7ae547b8af
"not merges" -> "not merged" that was wrong. "not merges" was
about flag --no-merges. Then several commits rephrased this
sentence a bit, but original line was right.

I've changed "not merges" -> "not merge commits" so that
this doesn't happen again.