Extensive tweaking of "Contributing to a Project" section
Changes include:

  - standardize use of " -- " for em dashes
  - small number of font changes
  - replace comments in code excerpts with ellipses
  - Substantial rewrite of sizable chunk of chapter

Rationale for some of the changes:

  First, I reworded phrases like "merged in", "pulled down" and "pushed up"
since those phrases are unnecessarily verbose; one can simply write "merged",
"pushed" and "pulled".

  Next, I replaced "comments" in some of the code excerpts with the same text
within ellipses since, in the middle of a series of commands, a like like this:

  # work

aesthetically looks like a command being run by root, so I replaced that with:

    ... work ...

which seems to be more visually obvious.

  Finally (and unintentionally), I started some simple rewording, but as I
reworded, those changes required further rewording to match the earlier
changes, which then required more rewording, and so on, and so on. It just
snowballed.