fixed minor grammar mistake
Guessing this should be from/to, as that's what a "trivial merge" would probably do:
so all of the files you _haven't_ changed in will be updated.   <=was
so all of the files you _haven't_ changed will be updated.       <=suggesting (as simpler for the other option below)

rather than from/to (which :
so all of the files you _haven't_ changed in will be updated.     <==was
so all of the files you _haven't_ changed in it will be updated.  <== possible alternative