Corrected the suggested changes from the PR Review
I really don't know what happened to the tags... maybe because I export it from the phone as a pdf and copy paste it to the PC. In case your wondering why I did that, I originally used my phone for the translation and I accidentally deleted all my progress and the latest good copy that I have is a backup of a text editor app. On top of that, if directly copy paste it from the editor there were weird characters appearing thus pdf was used. lol

Anyways, thank you for the review guys. 

P.S. From now on, I'll be working on a PC. haha