Add a section about pulling superproject updates
After a project which includes a submodule updates the submodule,
and you run 'git pull' in you local checkout of the superproject,
you then need to run 'git submodule update' to finalize the update
of your submodule checkout. This was not mentioned, so this commit
adds a section to that effect.