Add note: `git log` doesn't show upstream branches
I'm reading the book for the first time. I found something confusing so I've added what I hope is a helpful note.

I found it confusing when I tried to use `git log` to replicate what is shown in the graphic (named "HEAD moves when you checkout"); I thought I would see both the "testing" and "master" branches, like in the graphic, but was only able to see "master".