(#1395) Partial globbing is available in refspec since git 2.6.0
Motivation
The docs were saying that partial glob patterns are unavailable in
refspec, and namespacing is the only way to get flexible refspecs
matching multiple branches. This is no longer true since Git 2.6.0 which
added support for partial globbing patterns.

Modifications
The outdated paragraph was changed so that it now explicitly states that
the approach works.

Closes #1395.