[Fix] remove period mark from the figure captions
If a paragraph ends with a figure caption, it ends with double period
marks. One of them is due to figure caption, which itself ends with a
period and other is for end of paragraph.

So removing period from figure captions would be a good idea, as Ben has
suggested[1].

Due to this commit, now some paragraphs don't have any period mark at
end of paragraph. Those period are their because of caption, that's
paragraphs itself don't end with any period, see [2].

[1]: https://github.com/progit/progit2/issues/1437#issuecomment-643633534
[2]: https://github.com/progit/progit2/pull/1444#issuecomment-648754034

Resolves: #1437