Improve language
Correct usage of braces vs brackets
Rewrite first sentence to flow better, and to explain the braces situation a bit better.
Remove "either" from the second sentence, as I think the sentence is better without that word. The "or" in the middle of the sentence makes it clear that there are two choices to deal with the bracket issue. So "either" is not really adding value.

Co-authored-by: HonkingGoose <34918129+HonkingGoose@users.noreply.github.com>