Make wording in refs.asc clearer
When reading this line I was very confused.  It didn't sound like it made sense.  So I checked out a remote ref, ran `cat .git/HEAD`, and lo and behold it did "point" to the same thing that the remote branch pointed to.

I played around with it a little bit more and realized the difference between checking out a local and remote reference is that checking out a local reference sets `.git/HEAD` to a SYMBOLIC ref, while checking out a remote reference sets it to the exact commit.

I think the current language is *technically* correct, because "Git won't point HEAD at one" means it won't point HEAD to the remote reference itself (since "one" means the remote reference, not the commit).  So I'm not saying the current wording is bad, just that it confused me, so it might confuse others