Add '--nofork' flag to vim example in Appendix C
The -f/--nofork flag is suggested to be used "... when Vim is executed by a
program that will wait for the edit session to finish...." However, this flag
is used when Vim has the GUI version, effectively no-op'ing when non-GUI Vim.
This mirrors the existing "--nofork" flag for the Windows `gvim` example.