Mention includes in configuration section
When a user belongs to several organizations they may have different
identities or settings for the repos belonging to each
organization. Conditional include allow to apply those configuration
settings to each group without defining them on each repo
individually.

A big problem with the feature is its discoverability. A brief mention
in the configuration section to check out the details in the
documentation can dramatically raise awareness.