Correction recording-changes.asc
Modified the statement regarding the 'git diff' command to clarify that it compares the changes between the files in your working directory and the files in your Git repository, and it shows the differences between the two sources. By default, it displays the changes that are not yet staged for commit. However, if you use the '--staged' or '--cached' option, 'git diff' will compare the changes between the files in your staging area and the latest commit in your Git repository.