Update JGit links to new Eclipse project URLs
The previous JGit links lead to 404 page not found.

This was [reported over in the git-scm.com repository](https://github.com/git/git-scm.com/issues/2044), and to my utter disappointment the reporter not only did a shoddy job at reporting it there, not only a shoddy job of saying "I cannot find the correct location" instead of typing in "jgit" in the search page that is _linked directly from the 404 page_. And of course the reporter deemed it totally beneath himself to contribute the fix.

Sometimes I am not happy in the open source world. Most people there are nice. But not all of them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>